+++ frb_example/dart_minimal/rust/src/api/minimal.rs
@@ -8,26 +8,3 @@ pub fn init_app() {
 pub fn minimal_adder(a: i32, b: i32) -> i32 {
     a + b
 }
-
-#[frb]
-pub struct StructWithFieldRenameTwinNormal {
-    #[frb(name = "renamed_field")]
-    pub class: i32,
-}
-
-pub fn func_for_struct_with_field_rename_twin_normal(
-    arg: StructWithFieldRenameTwinNormal,
-) -> StructWithFieldRenameTwinNormal {
-    arg
-}
-
-pub struct StructWithDartKeywordFieldTwinNormal {
-    pub class: i32,
-    pub interface: i64,
-}
-
-pub fn func_for_struct_with_dart_keyword_field_twin_normal(
-    arg: StructWithDartKeywordFieldTwinNormal,
-) -> StructWithDartKeywordFieldTwinNormal {
-    arg
-}
