+++ frb_example/pure_dart/rust/src/api/structure.rs
@@ -1,3 +1,5 @@
+use flutter_rust_bridge::frb;
+
 pub struct StructWithZeroFieldTwinNormal {}
 
 pub fn func_struct_with_zero_field_twin_normal(
@@ -40,5 +42,28 @@ pub struct TupleStructWithTwoFieldTwinNormal(pub i32, pub i32);
 pub fn func_tuple_struct_with_two_field_twin_normal(
     arg: TupleStructWithTwoFieldTwinNormal,
 ) -> TupleStructWithTwoFieldTwinNormal {
+    arg
+}
+
+#[frb]
+pub struct StructWithFieldRenameTwinNormal {
+    #[frb(name = "renamed_field")]
+    pub class: i32,
+}
+
+pub fn func_for_struct_with_field_rename_twin_normal(
+    arg: StructWithFieldRenameTwinNormal,
+) -> StructWithFieldRenameTwinNormal {
+    arg
+}
+
+pub struct StructWithDartKeywordFieldTwinNormal {
+    pub class: i32,
+    pub interface: i64,
+}
+
+pub fn func_for_struct_with_dart_keyword_field_twin_normal(
+    arg: StructWithDartKeywordFieldTwinNormal,
+) -> StructWithDartKeywordFieldTwinNormal {
     arg
 }
